% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{add_axis}
\alias{add_axis}
\title{Add an axis to a timeline plot}
\usage{
add_axis(chart = NULL, scale = NULL, orient = NULL, ticks = NULL,
  tickValues = NULL, tickSize = NULL, innerTickSize = NULL,
  outerTickSize = NULL, tickPadding = NULL, tickFormat = NULL,
  styles = NULL)
}
\arguments{
\item{chart}{timeline plot htmlwidget to which you would like to add an axis}

\item{scale}{}

\item{orient}{}

\item{ticks}{}

\item{tickValues}{}

\item{tickSize}{}

\item{innerTickSize}{}

\item{outerTickSize}{}

\item{tickPadding}{}

\item{styles}{}
}
\value{
timeline chart modified with an added axis
}
\description{
Often, a timeline plot needs an x axis, so this little function supplies a very thin
wrapper for \href{https://github.com/mbostock/d3/wiki/SVG-Axes}{d3.svg.axis}, so that
we can add it to our timeline htmlwidget.
}

