% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{add_pretty_breaks_and_labels_to_one_oplot}
\alias{add_pretty_breaks_and_labels_to_one_oplot}
\title{If need arises to add xlabel, they can be added as prefix by passing to this function}
\usage{
add_pretty_breaks_and_labels_to_one_oplot(ggobject, prt_brks, xlabels)
}
\description{
If need arises to add xlabel, they can be added as prefix by passing to this function
}

