% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/components.R
\name{add_tasks}
\alias{add_tasks}
\title{Add post-render JavaScript tasks to timeline}
\usage{
add_tasks(chart, task = NULL)
}
\arguments{
\item{chart}{timeline plot htmlwidget}

\item{task}{either a \code{character} or \code{\link[htmlwidgets]{JS}}
         representing a JavaScript function to run after a timeline has
         been rendered.  This provides the ability for advanced customization.
         The JavaScript function will be \code{call}ed so \code{this.el} will be
         the containing \code{div} element and \code{this.chart} will be the
         \code{timeline} object.}
}
\description{
Add post-render JavaScript tasks to timeline
}

