% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_sed}
\alias{download_sed}
\title{download_sed}
\usage{
download_sed()
}
\value{
function returns a data frame with all observations contained in
the NOAA hurricane database
}
\description{
This function automatically downloads the database of hurricanes from the
National Centres for Environmental Information, compiled by the National
Oceanic and Atmospheric Administration (NOAA).
}
\examples{
all_hurricanes <- download_sed()

}
