% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{a data frame to be cleaned}
}
\value{
This function returns a cleaned data frame.
}
\description{
Takes the data frame downloaded from the NOAA and cleans it:
converts lattitute and longitude to numeric types, and creates a DATE
column.
}
\examples{
\donttest{all_hurricanes_clean <- eq_clean_data(all_hurricanes)}

}
