% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_mapping.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
eq_create_label
}
\examples{
usa_hurricanes\%>\%eq_create_label(.)\%>\%eq_map(annot_col = 'popup_text')

}
