% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{a data frame with LOCATION_DATA column to be cleaned.}
}
\value{
a data frame with extraneous data removed from column
LOCATION_NAME.
}
\description{
This function takes the NOAA hurricane data and cleans the LOCATION_NAME
column.
}
\examples{
\donttest{final_hurricanes <- eq_location_clean(all_hurricanes_clean)}

}
