% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_mapping.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{a data frame}

\item{annot_col}{column to take annotations from}
}
\value{
leaflet map
}
\description{
eq_map
}
\examples{
\donttest{eq_map(usa_hurricanes, annot_col = 'DATE')}

}
