% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{generate_color_mapping}
\alias{generate_color_mapping}
\title{Generate color mapping for all plots}
\usage{
generate_color_mapping(df, default_color_mapping)
}
\arguments{
\item{df}{Data frame to be plotted using plot_timeline}

\item{default_color_mapping}{Value to color mapping for all the plots in the data frame. ex: default_color_mapping <- c("0" = "#BCBEC0", "1" = "#1279C6")}
}
\description{
When all the plots have the same set of values, instead of typing the color mapping for all plots, default color to value mapping can be given which will generate the color mapping for all the plots based on the unique values present in each column
}
