% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "timeline",
  position = "identity", show.legend = TRUE, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{show.legend}{}

\item{na.rm}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
geom_timeline
}
\section{Aesthetics}{

geom_timeline understands the following aesthetics (required aesthetics 
are in bold):
\itemize{
 \item{\strong{"x"}}{column of longitude values}
 \item{\strong{"xmin"}}{column of latitude values}
 \item{"y"}{column of North West maximum wind radial extent, in 
 nautical miles}
 \item{"shape"}{not recommended}
 \item{"col"}{desired color, entered as a single string, or on column}
 \item{"fill"}{sets tranparency}
 \item{"alpha"}{not recommended}
}
}

\examples{
usa_hurricanes <- final_hurricanes\%>\%filter(COUNTRY == "USA")
ggplot(data = usa_hurricanes, aes(DATE, COUNTRY)) + geom_timeline(aes(xmin =as.Date("1999-01-01")))

}
