% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "timelabel",
  position = "identity", show.legend = FALSE, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{show.legend}{}

\item{na.rm}{}

\item{inherit.aes}{}

\item{...}{}
}
\value{

}
\description{
geom_timeline_label
}
\section{Aesthetics}{

geom_timeline_label understands the following aesthetics (required aesthetics 
are in bold):
\itemize{
 \item{\strong{"x"}}{column of longitude values}
 \item{\strong{"xmin"}}{column of latitude values}
 \item{\strong{"label"}}{column of North West maximum wind radial extent, in 
 nautical miles}
 \item{"n_max"}{not recommended}
 \item{"lwd"}{desired color, entered as a single string, or on column}
 \item{"col"}{sets tranparency}
 \item{"fontsize"}{not recommended}
}
}

\examples{
ggplot(data = usa_hurricanes, aes(DATE, COUNTRY)) + geom_timeline(aes(xmin =as.Date("1990-01-01"))) + geom_timeline_label(aes(xmin =as.Date("1990-01-01"), label = LOCATION_NAME))
}
