% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{rbind_grob_plots}
\alias{rbind_grob_plots}
\title{plots two or more ggplots one below another}
\usage{
rbind_grob_plots(all_plots_grob_scaled)
}
\arguments{
\item{all_plots}{is a list of ggplot or gtable objects}

\item{ratio}{if not null, is a numeric vector of the same length as all_plots. Plots are resized to this ratio.}
}
\description{
plots two or more ggplots one below another
}

