% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline(...)
}
\arguments{
\item{...}{additional element specifications not part of base ggplot2. If
supplied \code{validate} needs to be set to \code{FALSE}.}
}
\description{
dedicated theme for use with the timelineR package.
}
