% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_mapping.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{}

\item{annot_col}{}
}
\value{

}
\description{
eq_map
}
\examples{
eq_map(usa_hurricanes, annot_col = 'DATE')

}
