% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "timelabel",
  position = "identity", show.legend = FALSE, na.rm = FALSE,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_. 
If specified and inherit.aes = TRUE (the default), it is combined 
with the default mapping at the top level of the plot. You must 
supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three 
options:
If NULL, the default, the data is inherited from the plot data as 
specified in the call to ggplot.

A data.frame, or other object, will 
override the plot data. All objects will be fortified to produce 
a data frame. See fortify for which variables will be created. 

A function will be called with a single argument, the plot data. 
The return value must be a data.frame., and will be used as the 
layer data.}

\item{stat}{The statistical transformation to use on the data for 
this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the 
result of a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the 
legends? NA, the default, includes if any aesthetics are mapped. 
FALSE never includes, and TRUE always includes.}

\item{na.rm}{If FALSE, the default, missing values are removed 
with a warning. If TRUE, missing values are silently removed.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, 
rather than combining with them. This is most useful for helper 
functions that define both data and aesthetics and shouldn't inherit 
behaviour from the default plot specification, e.g. borders.}

\item{...}{other arguments passed on to layer. These are often aesthetics, 
used to set an aesthetic to a fixed value, like color = "red" or 
size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
gglist
}
\description{
Geom for labelling for geom_timeline
}
\section{Aesthetics}{

geom_timeline_label understands the following aesthetics (required aesthetics 
are in bold):
\itemize{
 \item{\strong{"x"}}{column of date values}
 \item{\strong{"xmin"}}{date to start plot, ending at last date}
 \item{\strong{"label"}}{column of labels to plot}
 \item{\strong{"n_max"}}{maximum number of labels to plot}
 \item{\strong{"n_max_arrange"}}{column used to organize, in descending order, n_max labels}
 \item{"lwd"}{desired color, entered as a single string, or on column}
 \item{"col"}{sets colour}
 \item{"fontsize"}{sets fontsize}
}
}

\examples{
\donttest{ggplot(data = usa_hurricanes, aes(DATE, COUNTRY)) + geom_timeline(aes(xmin =as.Date("1990-01-01"))) + geom_timeline_label(aes(xmin =as.Date("1990-01-01"), label = LOCATION_NAME))}

}
